Create canvasScene.js
canvasScene is a scene of Kellenberg, Casey, Public Square, Hagan, and that other one... something with an "S"? idk.
It currently has no mathematical scale, so there is no latitude/longitude calculations nor feet to pixels measurements.
It creates cubes from an array and randomly places them in the center of all the buildings, and assigns a random color to them.